Add module and update simplecov coverage
